pyproject.toml
[project]
name = "interactive-presentation"
version = "0.1.0"
description = ""
authors = [
    {name = "Viktor Jonsson",email = "viktor.jonsson@cellmax.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib (>=3.10.8,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "qrcode (>=8.1.1,<9.0.0) ; python_version >= \"3.13\" and python_version < \"4.0\"",
    "pillow (>=11.0.0,<12.0.0) ; python_version >= \"3.13\" and python_version < \"4.0\"",
    "field-viewer @ git+https://github.com/cellmaxtechnologies/packages.field-viewer.git"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
